feat: add types for sdk
