+++ package.json
@@ -4,6 +4,7 @@
 	"description": "zeroDAO Protocol SDK",
 	"main": "dist/lib/zero.js",
 	"repository": "https://github.com/zerodao-finance/zero-protocol",
+	"types": "./lib/types.d.ts",
 	"scripts": {
 		"test": "hardhat test",
 		"test:p2p:unit": "NODE_ENV=test yarn mocha -r ts-node/register lib/p2p/test/mockPubsub.test.ts lib/p2p/test/mockZero.test.ts --exit --timeout 0",
